+++ Gym-Of-Gyms/Models/Exercise.cs
@@ -4,7 +4,7 @@ public class Exercise
 {
     public int Exercise_Id { get; set; }
     public string Name { get; set; } = null!;
-    public double? Energy_Cost { get; set; }
-    public double? Height_Mass { get; set; }
+    public decimal? Energy_Cost { get; set; }
+    public decimal? Height_Mass { get; set; }
 
 }
