+++ Gym-Of-Gyms/Models/Food.cs
@@ -4,8 +4,8 @@ public class Food
 {
     public int Food_Id { get; set; }
     public string Name { get; set; } = null!;
-    public double Calories { get; set; }
-    public double Protein { get; set; }
-    public double Fat { get; set; }
-    public double Carbohydrates { get; set; }
+    public decimal Calories { get; set; }
+    public decimal Protein { get; set; }
+    public decimal Fat { get; set; }
+    public decimal Carbohydrates { get; set; }
 }
