+++ Gym-Of-Gyms/Models/Record_Food.cs
@@ -3,7 +3,7 @@
 public class Record_Food
 {
     public int Record_Food_Id { get; set; }
-    public double Mass { get; set; }
+    public decimal Mass { get; set; }
     public int Eating_Id { get; set; }
     public int Food_Id { get; set; }
     public Eating Eating { get; set; } = null!;
